Partway finished with the  KL tracker from the board
It's been two days on this island, and I believe I'm beginning to crack a little bit. Symptoms I am experiencing include confusion, nausea, and a feeling of a lack of meaning in my life. Withdrawal symptoms from human interaction are noticeable. I'll continue adding to the KL tracker as specified on the board to pass the time.